Refactor: Modularize map components and clean up GameMap
- Extracted `SaveGameplayAreaButton`, `SearchControl`, and `DrawControl` into separate modular components for better maintainability and reusability.
- Removed unused imports (`Input`, `SearchIcon`, `useMap`) and cleaned up the codebase.
- Updated the `GameMap` component to use the new modular components, improving readability and reducing complexity.
- Fixed a typo in the `SaveGameplayAreaButton` text ("araa" → "area").
- Ensured consistent code formatting and removed redundant code.
